requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
celery==5.4.0
redis==5.0.8
httpx==0.27.2
python-dotenv==1.0.1
prometheus-client==0.21.0
pyyaml==6.0.2
pytest==8.3.3
black==24.8.0
ruff==0.6.9
isort==5.13.2
reportlab==4.4.4
openpyxl==3.1.2
